Toast Updated

+++ app/Http/Controllers/CategoryController.php
@@ -6,7 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Models\Category;
 use Illuminate\Http\Request;
 
-class CategoryController extends Controller {
+class CategoryController extends Controller
+{
 
     public function __construct()
     {
@@ -16,11 +17,11 @@ class CategoryController extends Controller {
     public function get_index()
     {
         $categories = Category::all();
-        foreach($categories as $c){
-            if($c->cat_parent_id=='-1'){
+        foreach ($categories as $c) {
+            if ($c->cat_parent_id == '-1') {
                 $c->parent = '-';
-            }else{
-                $c->parent = Category::where('id',$c->cat_parent_id)->pluck('cat_title');
+            } else {
+                $c->parent = Category::where('id', $c->cat_parent_id)->pluck('cat_title');
             }
         }
         return view('admin.category.index')
@@ -29,7 +30,7 @@ class CategoryController extends Controller {
 
     public function get_create()
     {
-        $categories = Category::where('cat_parent_id','-1')->get();
+        $categories = Category::where('cat_parent_id', '-1')->get();
         return view('admin.category.create')
             ->with(compact('categories'));
     }
@@ -38,31 +39,39 @@ class CategoryController extends Controller {
     {
         Category::create($request->all());
         return redirect('/category')
-            ->with('success','Category created');
+            ->with('success', 'Category created');
     }
 
     public function get_edit($id)
     {
-        $categories = Category::where('cat_parent_id','-1')->get();
-        $category = Category::where('id',$id)->first();
+        $categories = Category::where('cat_parent_id', '-1')->get();
+        $category = Category::where('id', $id)->first();
         return view('admin.category.edit')
-            ->with(compact('category','categories'));
+            ->with(compact('category', 'categories'));
     }
 
-    public function post_update($id,Request $request)
+    public function post_update($id, Request $request)
     {
         $input = $request->all();
         unset($input['_token']);
-        Category::where('id',$id)->update($input);
+        Category::where('id', $id)->update($input);
         return redirect('/category')
-            ->with('success','Category updated');
+            ->with('success', 'Category updated');
     }
 
     public function get_delete($id)
     {
-        Category::destroy($id);
-        return redirect('/category')
-            ->with('success','Category deleted');
+        $check = Category::where('cat_parent_id', $id)->first();
+        if (empty($check)) {
+            Category::destroy($id);
+            return redirect('/category')
+                ->with('success', 'Category deleted');
+        } else {
+            return redirect('/category')
+                ->with('error', 'Please remove sub category first');
+        }
+
+
     }
 
 }
